requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-ckeditor==6.6.1
django-cors-headers==4.2.0
django-filter==23.2
django-jazzmin==2.6.0
django-modeltranslation==0.18.11
django-rosetta==0.9.9
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.5
Pillow==10.0.0
pymemcache==4.0.0
python-dotenv==1.0.0
whitenoise==6.5.0
python-telegram-bot==20.4
celery[librabbitmq]==5.3.1
django-phonenumber-field==7.1.0
phonenumbers==8.13.17



